refactor: move pdf and image to core tools (#262)
- Move src/plugins/media/{pdf,image}.ts → src/tools/{pdf,image}.ts
- Register in allTools and add to read/write/full tool scopes
- Drop tools/toolDescriptions from media plugin (plugin now owns only
  the attachment pipeline — save, digest, resolve, HTTP serving)
- Add pdf and image to SUBAGENT_TOOLS allowlist and system prompt

No behavior change for existing agents. Sub-agents gain access to
pdf and image. Any surface that gates by toolScope (sub-agents,
plugin-less setups) can reach them without plugin initialization.
