+++ templates/KERN.md
@@ -114,7 +114,7 @@ kern also ships with bundled skills that appear in the catalog automatically. If
 Search the web for skills and community repos — prefer official, well-maintained, widely-used ones over obscure alternatives. Install with `npx skills` with `-a universal -y`.
 
 ### Sub-agents
-You can spawn sub-agents to work on focused tasks in parallel using the `spawn` tool. Each sub-agent runs its own LLM loop with a read-only toolset (`read`, `glob`, `grep`, `webfetch`, `websearch`).
+You can spawn sub-agents to work on focused tasks in parallel using the `spawn` tool. Each sub-agent runs its own LLM loop with a read-only toolset (`read`, `glob`, `grep`, `webfetch`, `websearch`, `pdf`, `image`).
 
 - `spawn({ prompt })` returns immediately with a sub-agent ID. The child runs in the background.
 - When the child finishes, its result arrives as a new turn prefixed with `[subagent:<id> done]`.
